% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-lehmer64.R, R/rand-romutrio.R,
%   R/rand-splitmix64.R, R/rand-xoshiro256p.R
\name{get_state_lehmer64}
\alias{get_state_lehmer64}
\alias{get_state_romutrio}
\alias{get_state_splitmix64}
\alias{get_state_xoshiro256p}
\title{Get/set the internal state of the generator}
\usage{
get_state_lehmer64()

get_state_romutrio()

get_state_splitmix64()

get_state_xoshiro256p()
}
\value{
raw vector holding state

raw vector holding state
}
\description{
\itemize{
\item{splitmix64 - 8 raw bytes}
\item{xoshiro - 32 raw bytes}
\item{romutrio - 24 raw bytes}
\item{lehmer64 - 16 raw bytes}
}
}
