% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-xoshiro256p.R
\name{jump_xoshiro256p}
\alias{jump_xoshiro256p}
\title{This is the jump function for the generator. It is equivalent
to 2^128 calls to next(); it can be used to generate 2^128
non-overlapping subsequences for parallel computations.}
\usage{
jump_xoshiro256p()
}
\description{
This is the jump function for the generator. It is equivalent
to 2^128 calls to next(); it can be used to generate 2^128
non-overlapping subsequences for parallel computations.
}
