% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-xoshiro256p.R
\name{long_jump_xoshiro256p}
\alias{long_jump_xoshiro256p}
\title{This is the long-jump function for the generator. It is equivalent to
2^192 calls to next(); it can be used to generate 2^64 starting points,
from each of which jump() will generate 2^64 non-overlapping
subsequences for parallel distributed computations.}
\usage{
long_jump_xoshiro256p()
}
\description{
This is the long-jump function for the generator. It is equivalent to
2^192 calls to next(); it can be used to generate 2^64 starting points,
from each of which jump() will generate 2^64 non-overlapping
subsequences for parallel distributed computations.
}
