% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-lehmer64.R, R/rand-romutrio.R,
%   R/rand-splitmix64.R, R/rand-xoshiro256p.R
\name{runif1_lehmer64}
\alias{runif1_lehmer64}
\alias{runif1_romutrio}
\alias{runif1_splitmix64}
\alias{runif1_xoshiro256p}
\title{Fetch single uniform random number}
\usage{
runif1_lehmer64()

runif1_romutrio()

runif1_splitmix64()

runif1_xoshiro256p()
}
\description{
Fetch single uniform random number
}
