% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-lehmer64.R, R/rand-romutrio.R,
%   R/rand-splitmix64.R, R/rand-xoshiro256p.R
\name{runif_lehmer64}
\alias{runif_lehmer64}
\alias{runif_romutrio}
\alias{runif_splitmix64}
\alias{runif_splitmix64_as_bytes}
\alias{runif_xoshiro256p}
\title{Generate a vector of uniform random numbers}
\usage{
runif_lehmer64(n, min = 0, max = 1)

runif_romutrio(n, min = 0, max = 1)

runif_splitmix64(n, min = 0, max = 1)

runif_splitmix64_as_bytes(n, min = 0, max = 1)

runif_xoshiro256p(n, min = 0, max = 1)
}
\arguments{
\item{n}{Number of observations}

\item{min, max}{lower and upper limit.}
}
\description{
Generate a vector of uniform random numbers
}
