% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-lehmer64.R, R/rand-romutrio.R,
%   R/rand-splitmix64.R, R/rand-xoshiro256p.R
\name{set_seed_lehmer64}
\alias{set_seed_lehmer64}
\alias{set_seed_romutrio}
\alias{set_seed_splitmix64}
\alias{set_seed_xoshiro256p}
\title{Set PRNG seed}
\usage{
set_seed_lehmer64(seedlo, seedhi = 524287L)

set_seed_romutrio(seedlo, seedhi = 524287L)

set_seed_splitmix64(seedlo, seedhi = 524287L)

set_seed_xoshiro256p(seedlo, seedhi = 524287L)
}
\arguments{
\item{seedlo, seedhi}{PRNG seeds need to be 64-bit unsigned integers - which
aren't actually a type in R.  Instead allow user to pass in 2 32bit
integers - \code{seedhi} will be shifted by 32bits to be the upper
part of a reconstructed 64bit int within the C code.  Default for seedhi is
524287L (6th mersenne prime) - no real justification for this choice.
I think this counts as a Nothing-Up-My-Sleeve number.}
}
\description{
Set PRNG seed
}
