% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-lehmer64.R, R/rand-romutrio.R,
%   R/rand-splitmix64.R, R/rand-xoshiro256p.R
\name{set_state_lehmer64}
\alias{set_state_lehmer64}
\alias{set_state_romutrio}
\alias{set_state_splitmix64}
\alias{set_state_xoshiro256p}
\title{Set the internal state of generator}
\usage{
set_state_lehmer64(state)

set_state_romutrio(state)

set_state_splitmix64(state)

set_state_xoshiro256p(state)
}
\arguments{
\item{state}{raw vector of length 8}
}
\description{
\itemize{
\item{splitmix64 - 8 raw bytes}
\item{xoshiro - 32 raw bytes}
\item{romutrio - 24 raw bytes}
\item{lehmer64 - 16 raw bytes}
}
}
